% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_intensity_area.R
\name{calc_intensity_area}
\alias{calc_intensity_area}
\title{Calculate carbon footprint intensity per hectare}
\usage{
calc_intensity_area(total_emissions, area_ha)
}
\arguments{
\item{total_emissions}{Numeric. Total emissions in kg CO2eq (from \code{calc_total_emissions()}).}

\item{area_ha}{Numeric. Total land area in hectares used by the dairy farm.}
}
\value{
Numeric. Emissions intensity (kg CO2eq/ha).
}
\description{
Computes emissions intensity as kg CO2eq per hectare of utilized land.
}
\examples{
b <- set_system_boundaries("farm_gate")
e1 <- calc_emissions_entero(100, boundaries = b)
e2 <- calc_emissions_manure(100, boundaries = b)
tot <- calc_total_emissions(e1, e2)
calc_intensity_area(tot$total_co2eq, area_ha = 120)
}
