% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_template.R
\name{download_template}
\alias{download_template}
\title{Download cowfootR Excel template}
\usage{
download_template(file = "cowfootR_template.xlsx")
}
\arguments{
\item{file}{Path where the template will be saved. Default = "cowfootR_template.xlsx".}
}
\value{
Invisibly returns the file path.
}
\description{
Saves a blank Excel template with required columns for batch carbon footprint calculations.
}
\examples{
download_template("my_farms.xlsx")
}
