% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_batch.R
\name{export_hdc_report}
\alias{export_hdc_report}
\title{Export cowfootR batch results to Excel}
\usage{
export_hdc_report(
  batch_results,
  file = "cowfootR_report.xlsx",
  include_details = FALSE
)
}
\arguments{
\item{batch_results}{A cf_batch_complete object returned by \code{calc_batch()}.}

\item{file}{Path to the Excel file to save. Default = "cowfootR_report.xlsx".}

\item{include_details}{Logical. If TRUE, includes extra sheets with detailed objects (if available).}
}
\value{
Invisibly returns the file path.
}
\description{
Exports results from \code{calc_batch()} into an Excel file
with summary and farm-level sheets.
}
