% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_hdc.R
\name{report_hdc}
\alias{report_hdc}
\title{Generate a carbon footprint report}
\usage{
report_hdc(total_emissions, intensity_litre, intensity_area)
}
\arguments{
\item{total_emissions}{List. Output of \code{calc_total_emissions()}.}

\item{intensity_litre}{Numeric. Output of \code{calc_intensity_litre()}.}

\item{intensity_area}{Numeric. Output of \code{calc_intensity_area()}.}
}
\value{
A list of gt tables: breakdown and summary.
}
\description{
Creates a summary table of carbon footprint results,
including breakdown by source and intensities.
}
\examples{
b <- set_system_boundaries("farm_gate")
e1 <- calc_emissions_entero(100, boundaries = b)
e2 <- calc_emissions_manure(100, boundaries = b)
tot <- calc_total_emissions(e1, e2)
lit <- calc_intensity_litre(tot$total_co2eq, 750000)
are <- calc_intensity_area(tot$total_co2eq, 120)
report_hdc(tot, lit, are)
}
