% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_hdc_report.R
\name{export_hdc_report}
\alias{export_hdc_report}
\title{Export carbon footprint results to Excel}
\usage{
export_hdc_report(results, file = "HdC_results.xlsx")
}
\arguments{
\item{results}{Data.frame. Output from \code{calc_emissions_batch()}.}

\item{file}{Character. Path to the Excel file to be created.}
}
\value{
Invisibly returns the file path.
}
\description{
Saves the results from \code{calc_emissions_batch()} into an Excel file.
}
\examples{
# results <- calc_emissions_batch("cowfootR_template.xlsx")
# export_hdc_report(results, "HdC_results.xlsx")
}
