% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_intensity_area.R
\name{print.cf_area_intensity}
\alias{print.cf_area_intensity}
\title{Print method for cf_area_intensity objects}
\usage{
\method{print}{cf_area_intensity}(x, ...)
}
\arguments{
\item{x}{A cf_area_intensity object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cf_area_intensity objects
}
