% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_intensity_area.R
\name{print.cf_area_intensity}
\alias{print.cf_area_intensity}
\title{Print method for cf_area_intensity objects}
\usage{
\method{print}{cf_area_intensity}(x, ...)
}
\arguments{
\item{x}{A cf_area_intensity object}

\item{...}{Additional arguments (ignored)}
}
\value{
No return value, called for side effects. Prints formatted area intensity
information to the console and invisibly returns the input object.

The input object \code{x}, invisibly.
}
\description{
Print method for cf_area_intensity objects
}
\examples{
x <- list(
  intensity_per_total_ha = 900,
  intensity_per_productive_ha = 1100,
  land_use_efficiency = 0.92,
  total_emissions_co2eq = 108000,
  area_total_ha = 120,
  area_productive_ha = 110,
  date = Sys.Date()
)
class(x) <- "cf_area_intensity"
print(x)
}
