% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_intensity_litre.R
\name{print.cf_intensity}
\alias{print.cf_intensity}
\title{Print method for cf_intensity objects}
\usage{
\method{print}{cf_intensity}(x, ...)
}
\arguments{
\item{x}{A cf_intensity object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cf_intensity objects
}
