% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_total_emissions.R
\name{print.cf_total}
\alias{print.cf_total}
\title{Print method for cf_total objects}
\usage{
\method{print}{cf_total}(x, ...)
}
\arguments{
\item{x}{A cf_total object}

\item{...}{Additional arguments passed to print methods (currently ignored)}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for cf_total objects
}
